# public/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dastyare Social — SH Links API",
    "version": "1.0.0",
    "description": "REST API for short-link management. All endpoints require Bearer API key authentication via the `Authorization` header."
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Static API key configured via the `API_KEY` environment variable. Required for all API calls."
      }
    },
    "schemas": {
      "CreateLinkBody": {
        "type": "object",
        "properties": {
          "r_to": {
            "type": "string",
            "format": "uri",
            "description": "Destination URL"
          },
          "r_path": {
            "type": "string",
            "minLength": 2,
            "maxLength": 32,
            "pattern": "^[a-z0-9-]+$",
            "description": "Optional custom slug (2–32 lowercase alphanumeric/hyphen chars)"
          }
        },
        "required": ["r_to"],
        "description": "Create link request payload",
        "example": {
          "r_to": "https://workshop.dastyare.social/summer-2026",
          "r_path": "summer-2026"
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message"
          }
        },
        "required": ["error"],
        "description": "Error response",
        "example": {
          "error": "Link not found"
        }
      },
      "LinkListResponse": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/LinkResponse"
        },
        "description": "List of links",
        "example": [
          {
            "id": "a1b2c3d4e5f6g7h8",
            "r_path": "summer-2026",
            "r_to": "https://workshop.dastyare.social/summer-2026",
            "is_active": true,
            "redirects": "236",
            "createdAt": "2026-06-01T10:30:00.000Z",
            "updatedAt": "2026-07-15T14:20:00.000Z"
          },
          {
            "id": "b2c3d4e5f6g7h8i9",
            "r_path": "docs",
            "r_to": "https://docs.dastyare.social",
            "is_active": true,
            "redirects": "42",
            "createdAt": "2026-05-14T15:45:00.000Z",
            "updatedAt": "2026-05-14T15:45:00.000Z"
          }
        ]
      },
      "LinkPathParam": {
        "type": "object",
        "properties": {
          "link": {
            "type": "string",
            "description": "Link ID or r_path slug",
            "example": "summer-2026"
          }
        },
        "required": ["link"],
        "description": "Link path parameter",
        "example": {
          "link": "summer-2026"
        }
      },
      "LinkResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "Link ID"
          },
          "r_path": {
            "type": "string",
            "description": "Short path slug"
          },
          "r_to": {
            "type": "string",
            "description": "Destination URL"
          },
          "is_active": {
            "type": "boolean",
            "description": "Active state"
          },
          "redirects": {
            "type": "string",
            "description": "Redirect count"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "description": "Creation timestamp"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "description": "Last update timestamp"
          }
        },
        "required": [
          "id",
          "r_path",
          "r_to",
          "is_active",
          "redirects",
          "createdAt",
          "updatedAt"
        ],
        "description": "Link object",
        "example": {
          "id": "a1b2c3d4e5f6g7h8",
          "r_path": "summer-2026",
          "r_to": "https://workshop.dastyare.social/summer-2026",
          "is_active": true,
          "redirects": "236",
          "createdAt": "2026-06-01T10:30:00.000Z",
          "updatedAt": "2026-07-15T14:20:00.000Z"
        }
      },
      "SuccessResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Operation success status"
          }
        },
        "required": ["success"],
        "description": "Success response",
        "example": {
          "success": true
        }
      },
      "UpdateLinkBody": {
        "type": "object",
        "properties": {
          "r_to": {
            "type": "string",
            "format": "uri",
            "description": "New destination URL"
          },
          "is_active": {
            "type": "boolean",
            "description": "Active state (true = enabled, false = disabled)"
          }
        },
        "description": "Update link request payload",
        "example": {
          "r_to": "https://workshop.dastyare.social/winter-2026",
          "is_active": false
        }
      }
    },
    "responses": {}
  },
  "includePaths": ["/src/app/api/links/**", "/src/app/api/links/[link]/*"],
  "excludePaths": [
    "/src/app/api/auth/**",
    "/src/app/api/trpc/**",
    "/src/app/api/account/**"
  ],
  "tags": [
    {
      "name": "Links",
      "description": "Create, read, update, and delete short links."
    }
  ],
  "paths": {
    "/links": {
      "get": {
        "operationId": "get-links",
        "summary": "List all short links",
        "description": "Returns every short link ordered by creation date (newest first). Includes the redirect counter for each link.",
        "tags": ["Links"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Links retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkListResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "post-links",
        "summary": "Create a short link",
        "description": "Creates a new short link. A random slug is generated when `r_path` is omitted.",
        "tags": ["Links"],
        "parameters": [],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateLinkBody"
              }
            }
          },
          "description": "Create link request"
        },
        "responses": {
          "201": {
            "description": "Link created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — invalid URL or r_path format",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409": {
            "description": "Path already in use",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/links/{link}": {
      "get": {
        "operationId": "get-links-{link}",
        "summary": "Get a link by ID or slug",
        "description": "Fetches a single link by its ID or r_path slug (searches by ID first, then by slug).",
        "tags": ["Links"],
        "parameters": [
          {
            "in": "path",
            "name": "link",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Link ID or r_path slug"
            },
            "description": "Link ID or r_path slug",
            "example": "summer-2026"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Link retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "patch-links-{link}",
        "summary": "Update a link",
        "description": "Update `r_to` and/or `is_active` for a link identified by ID or slug.",
        "tags": ["Links"],
        "parameters": [
          {
            "in": "path",
            "name": "link",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Link ID or r_path slug"
            },
            "description": "Link ID or r_path slug",
            "example": "summer-2026"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateLinkBody"
              }
            }
          },
          "description": "Update link request"
        },
        "responses": {
          "200": {
            "description": "Link updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LinkResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error — invalid URL or no fields provided",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete-links-{link}",
        "summary": "Delete a link",
        "description": "Permanently deletes the link identified by ID or slug.",
        "tags": ["Links"],
        "parameters": [
          {
            "in": "path",
            "name": "link",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Link ID or r_path slug"
            },
            "description": "Link ID or r_path slug",
            "example": "summer-2026"
          }
        ],
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Link deleted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SuccessResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized — invalid or missing API key",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "Link not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "servers": [
    {
      "url": "/api",
      "description": "API server"
    }
  ]
}